+++ src/MediatRItemExtension.Shared/Enums/Codes/ReqInfoCodeType.cs
@@ -28,6 +28,7 @@ namespace MediatRItemExtension.Enums.Codes
         RF0001,
         RF0002,
         RF0003,
-        RF0004
+        RF0004,
+        RF0005
     }
 }